sysctl takes a 'size_t' variable for the size parameter
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V5-4-patches@19830 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/mibgroup/ucd-snmp/proc.c
@@ -360,8 +360,8 @@ get_proc_instance(struct myproc *proc, oid inst)
  * these are for keeping track of the proc array 
  */
 
-static int      nproc = 0;
-static int      onproc = -1;
+static size_t   nproc = 0;
+static size_t   onproc = -1;
 static struct kinfo_proc *pbase = 0;
 
 int
